+++ Chartboost/ChartboostBannerCustomEvent.h
@@ -1,0 +1,18 @@
+//
+//  ChartboostBannerCustomEvent.h
+//  MoPubSDK
+//
+//  Copyright (c) 2019 MoPub. All rights reserved.
+//
+
+#if __has_include(<MoPub/MoPub.h>)
+    #import <MoPub/MoPub.h>
+#elif __has_include(<MoPubSDKFramework/MoPub.h>)
+    #import <MoPubSDKFramework/MoPub.h>
+#else
+    #import "MPBannerCustomEvent.h"
+#endif
+
+@interface ChartboostBannerCustomEvent : MPBannerCustomEvent
+
+@end