+++ Chartboost/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 8.0.1.0
+      * This version of the adapters has been certified with Chartboost 8.0.1.
+      * Added banner adapter.
+
   * 7.5.0.0
       * This version of the adapters has been certified with Chartboost 7.5.0.
       * Fix empty Chartboost location Strings.